AG-17996 - remove unused trial licence form variants
Only the allFields variant was ever rendered: no call site passed a
`type` to the TrialLicenceForm dispatcher, so DEFAULT_TRIAL_LICENSE_FORM
('allFields') always won and the emailOnly/original variants were dead.

Promote the allFields implementation to TrialLicenceForm and drop the
dispatcher along with the now-unused TrialLicenceFormType and
DEFAULT_TRIAL_LICENSE_FORM constants.
